+++ lib/router.js
@@ -116,12 +116,12 @@ router.get('/mzitu/category/:category', lazyloadRouteHandler('./routes/mzitu/cat
 router.get('/mzitu/post/:id', lazyloadRouteHandler('./routes/mzitu/post'));
 router.get('/mzitu/tag/:tag', lazyloadRouteHandler('./routes/mzitu/tag'));
 
-// pixiv
-router.get('/pixiv/user/bookmarks/:id', lazyloadRouteHandler('./routes/pixiv/bookmarks'));
-router.get('/pixiv/user/illustfollows', lazyloadRouteHandler('./routes/pixiv/illustfollow'));
-router.get('/pixiv/user/:id', lazyloadRouteHandler('./routes/pixiv/user'));
-router.get('/pixiv/ranking/:mode/:date?', lazyloadRouteHandler('./routes/pixiv/ranking'));
-router.get('/pixiv/search/:keyword/:order?/:mode?', lazyloadRouteHandler('./routes/pixiv/search'));
+// pixiv migrated to v2
+// router.get('/pixiv/user/bookmarks/:id', lazyloadRouteHandler('./routes/pixiv/bookmarks'));
+// router.get('/pixiv/user/illustfollows', lazyloadRouteHandler('./routes/pixiv/illustfollow'));
+// router.get('/pixiv/user/:id', lazyloadRouteHandler('./routes/pixiv/user'));
+// router.get('/pixiv/ranking/:mode/:date?', lazyloadRouteHandler('./routes/pixiv/ranking'));
+// router.get('/pixiv/search/:keyword/:order?/:mode?', lazyloadRouteHandler('./routes/pixiv/search'));
 
 // pixiv-fanbox
 router.get('/fanbox/:user?', lazyloadRouteHandler('./routes/fanbox/main'));
