+++ lib/v2/pixiv/api/getBookmarks.js
@@ -3,23 +3,21 @@ const maskHeader = require('../constants').maskHeader;
 const queryString = require('query-string');
 
 /**
- * 按热门排序搜索内容
- * @param {string} keyword 关键词
+ * 获取用户的收藏
+ *
+ * @param {string} user_id 目标用户id
  * @param {string} token pixiv oauth token
  * @returns {Promise<got.AxiosResponse<{illusts: illust[]}>>}
  */
-module.exports = function searchPopularIllust(keyword, token) {
-    return got({
-        method: 'get',
-        url: 'https://app-api.pixiv.net/v1/search/popular-preview/illust',
+module.exports = function getBookmarks(user_id, token) {
+    return got('https://app-api.pixiv.net/v1/user/bookmarks/illust', {
         headers: {
             ...maskHeader,
             Authorization: 'Bearer ' + token,
         },
         searchParams: queryString.stringify({
-            word: keyword,
-            search_target: 'partial_match_for_tags',
-            filter: 'for_ios',
+            user_id,
+            restrict: 'public',
         }),
     });
 };
