+++ lib/v2/pixiv/api/getIllustFollows.js
@@ -3,22 +3,18 @@ const maskHeader = require('../constants').maskHeader;
 const queryString = require('query-string');
 
 /**
- * 获取用户插画作品
- * @param {string} user_id 目标用户id
+ * 获取用户关注的画师们的最新插画
  * @param {string} token pixiv oauth token
  * @returns {Promise<got.AxiosResponse<{illusts: illust[]}>>}
  */
-module.exports = function getIllusts(user_id, token) {
-    return got({
-        method: 'get',
-        url: 'https://app-api.pixiv.net/v1/user/illusts',
+module.exports = function getUserIllustFollows(token) {
+    return got('https://app-api.pixiv.net/v2/illust/follow', {
         headers: {
             ...maskHeader,
             Authorization: 'Bearer ' + token,
         },
         searchParams: queryString.stringify({
-            user_id,
-            filter: 'for_ios',
+            restrict: 'public',
         }),
     });
 };
