+++ lib/v2/pixiv/api/getRanking.js
@@ -14,9 +14,7 @@ const allowMode = ['day', 'week', 'month', 'day_male', 'day_female', 'week_origi
  */
 module.exports = function getRanking(mode, date, token) {
     assert(allowMode.includes(mode), 'Mode not allow.');
-    return got({
-        method: 'get',
-        url: 'https://app-api.pixiv.net/v1/illust/ranking',
+    return got('https://app-api.pixiv.net/v1/illust/ranking', {
         headers: {
             ...maskHeader,
             Authorization: 'Bearer ' + token,
