+++ lib/v2/pixiv/api/getUserDetail.js
@@ -15,9 +15,7 @@ const queryString = require('query-string');
  * @returns {Promise<got.AxiosResponse<userDetail>>}
  */
 module.exports = function getUserDetail(user_id, token) {
-    return got({
-        method: 'get',
-        url: 'https://app-api.pixiv.net/v1/user/detail',
+    return got('https://app-api.pixiv.net/v1/user/detail', {
         headers: {
             ...maskHeader,
             Authorization: 'Bearer ' + token,
