+++ lib/v2/pixiv/api/searchIllust.js
@@ -3,23 +3,22 @@ const maskHeader = require('../constants').maskHeader;
 const queryString = require('query-string');
 
 /**
- * 获取用户的收藏
- *
- * @param {string} user_id 目标用户id
+ * 按时间排序搜索内容
+ * @param {string} keyword 关键词
  * @param {string} token pixiv oauth token
  * @returns {Promise<got.AxiosResponse<{illusts: illust[]}>>}
  */
-module.exports = function getBookmarks(user_id, token) {
-    return got({
-        method: 'get',
-        url: 'https://app-api.pixiv.net/v1/user/bookmarks/illust',
+module.exports = function searchIllust(keyword, token) {
+    return got('https://app-api.pixiv.net/v1/search/illust', {
         headers: {
             ...maskHeader,
             Authorization: 'Bearer ' + token,
         },
         searchParams: queryString.stringify({
-            user_id,
-            restrict: 'public',
+            word: keyword,
+            search_target: 'partial_match_for_tags',
+            sort: 'date_desc',
+            filter: 'for_ios',
         }),
     });
 };
