+++ lib/v2/pixiv/api/searchPopularIllust.js
@@ -3,15 +3,13 @@ const maskHeader = require('../constants').maskHeader;
 const queryString = require('query-string');
 
 /**
- * 按时间排序搜索内容
+ * 按热门排序搜索内容
  * @param {string} keyword 关键词
  * @param {string} token pixiv oauth token
  * @returns {Promise<got.AxiosResponse<{illusts: illust[]}>>}
  */
-module.exports = function searchIllust(keyword, token) {
-    return got({
-        method: 'get',
-        url: 'https://app-api.pixiv.net/v1/search/illust',
+module.exports = function searchPopularIllust(keyword, token) {
+    return got('https://app-api.pixiv.net/v1/search/popular-preview/illust', {
         headers: {
             ...maskHeader,
             Authorization: 'Bearer ' + token,
@@ -19,7 +17,6 @@ module.exports = function searchIllust(keyword, token) {
         searchParams: queryString.stringify({
             word: keyword,
             search_target: 'partial_match_for_tags',
-            sort: 'date_desc',
             filter: 'for_ios',
         }),
     });
