+++ lib/v2/pixiv/constants.js
