+++ lib/v2/pixiv/maintainer.js
@@ -1,0 +1,7 @@
+module.exports = {
+    '/ranking/:mode/:date?': ['EYHN'],
+    '/search/:keyword/:order?/:mode?': ['DIYgod'],
+    '/user/bookmarks/:id': ['EYHN'],
+    '/user/illustfollows': ['ClarkeCheng'],
+    '/user/:id': ['DIYgod'],
+};
