+++ lib/v2/pixiv/pixiv-got.js
@@ -6,7 +6,7 @@ const pixivConfig = require('@/config').value.pixiv;
 
 async function dohResolve(name, jsonDohEndpoint) {
     try {
-        const response = await got.get(jsonDohEndpoint, {
+        const response = await got(jsonDohEndpoint, {
             searchParams: {
                 name,
                 type: 'A',
