+++ lib/v2/pixiv/ranking.js
@@ -1,7 +1,8 @@
-const getToken = require('./token');
+const { getToken } = require('./token');
 const getRanking = require('./api/getRanking');
 const config = require('@/config').value;
 const pixivUtils = require('./utils');
+const { parseDate } = require('@/utils/parse-date');
 
 const titles = {
     day: 'pixiv 日排行',
@@ -56,7 +57,7 @@ module.exports = async (ctx) => {
     const mode = alias[ctx.params.mode] ? alias[ctx.params.mode] : ctx.params.mode;
     const date = ctx.params.date ? new Date(ctx.params.date) : new Date();
 
-    const token = await getToken();
+    const token = await getToken(ctx.cache.tryGet);
     if (!token) {
         throw 'pixiv not login';
     }
@@ -75,7 +76,7 @@ module.exports = async (ctx) => {
             const images = pixivUtils.getImgs(illust);
             return {
                 title: `#${index + 1} ${illust.title}`,
-                pubDate: new Date(illust.create_date).toUTCString(),
+                pubDate: parseDate(illust.create_date),
                 description: `<p>画师：${illust.user.name} - 阅览数：${illust.total_view} - 收藏数：${illust.total_bookmarks}</p><br>${images.join('')}`,
                 link: `https://www.pixiv.net/artworks/${illust.id}`,
                 author: illust.user.name,
