+++ lib/v2/pixiv/search.js
@@ -1,8 +1,9 @@
-const getToken = require('./token');
+const { getToken } = require('./token');
 const searchPopularIllust = require('./api/searchPopularIllust');
 const searchIllust = require('./api/searchIllust');
 const config = require('@/config').value;
 const pixivUtils = require('./utils');
+const { parseDate } = require('@/utils/parse-date');
 
 module.exports = async (ctx) => {
     if (!config.pixiv || !config.pixiv.refreshToken) {
@@ -13,7 +14,7 @@ module.exports = async (ctx) => {
     const order = ctx.params.order || 'date';
     const mode = ctx.params.mode;
 
-    const token = await getToken();
+    const token = await getToken(ctx.cache.tryGet);
     if (!token) {
         throw 'pixiv not login';
     }
@@ -40,7 +41,7 @@ module.exports = async (ctx) => {
             return {
                 title: illust.title,
                 author: illust.user.name,
-                pubDate: new Date(illust.create_date).toUTCString(),
+                pubDate: parseDate(illust.create_date),
                 description: `<p>画师：${illust.user.name} - 阅览数：${illust.total_view} - 收藏数：${illust.total_bookmarks}</p>${images.join('')}`,
                 link: `https://www.pixiv.net/artworks/${illust.id}`,
             };
