+++ lib/v2/pixiv/user.js
@@ -1,7 +1,8 @@
-const getToken = require('./token');
+const { getToken } = require('./token');
 const getIllusts = require('./api/getIllusts');
 const config = require('@/config').value;
 const pixivUtils = require('./utils');
+const { parseDate } = require('@/utils/parse-date');
 
 module.exports = async (ctx) => {
     if (!config.pixiv || !config.pixiv.refreshToken) {
@@ -9,7 +10,7 @@ module.exports = async (ctx) => {
     }
 
     const id = ctx.params.id;
-    const token = await getToken();
+    const token = await getToken(ctx.cache.tryGet);
     if (!token) {
         throw 'pixiv not login';
     }
@@ -28,7 +29,7 @@ module.exports = async (ctx) => {
             return {
                 title: illust.title,
                 author: username,
-                pubDate: new Date(illust.create_date).toUTCString(),
+                pubDate: parseDate(illust.create_date),
                 description: `<p>画师：${username} - 阅览数：${illust.total_view} - 收藏数：${illust.total_bookmarks}</p>${images.join('')}`,
                 link: `https://www.pixiv.net/artworks/${illust.id}`,
             };
